requirements.txt
discord.py
feedparser
python-dotenv
aiofiles
requests
beautifulsoup4